+++ remote/src/main/java/com/khoshnaw/remote/api/MovieApi.kt
@@ -1,6 +1,6 @@
 package com.khoshnaw.remote.api
 
-import com.khoshnaw.remote.movie.GetMoviesResponse
+import com.khoshnaw.remote.apiDataSource.movie.GetMoviesResponse
 import retrofit2.Response
 import retrofit2.http.GET
