+++ remote/src/main/java/com/khoshnaw/remote/apiDataSource/base/APIDataSource.kt
@@ -1,0 +1,3 @@
+package com.khoshnaw.remote.apiDataSource.base
+
+abstract class APIDataSource
