+++ remote/src/main/java/com/khoshnaw/remote/apiDataSource/movie/MovieAPIDataSource.kt
@@ -1,15 +1,16 @@
-package com.khoshnaw.remote.movie
+package com.khoshnaw.remote.apiDataSource.movie
 
 import com.khoshnaw.entity.Movie
-import com.khoshnaw.remote.api.MovieApi
-import com.khoshnaw.remote.util.bodyOrException
 import com.khoshnaw.gateway.remoteDataSource.MovieRemoteDataSource
+import com.khoshnaw.remote.api.MovieApi
+import com.khoshnaw.remote.apiDataSource.base.APIDataSource
 import com.khoshnaw.remote.mapper.toEntity
+import com.khoshnaw.remote.util.bodyOrException
 import javax.inject.Inject
 
 class MovieAPIDataSource @Inject constructor(
     private val movieApi: MovieApi
-) : MovieRemoteDataSource {
+) : APIDataSource(), MovieRemoteDataSource {
     override suspend fun loadMovieList(): List<Movie> =
         movieApi.loadMovieList().bodyOrException().movieList.toEntity()
 }
