+++ remote/src/main/java/com/khoshnaw/remote/module/TMDBApiKey.kt
@@ -1,4 +1,4 @@
-package com.khoshnaw.remote.authentication
+package com.khoshnaw.remote.module
 
 data class TMDBApiKey(
     val apiKey: String
